pyproject.toml
[tool.poetry]
name = "animals-classifier"
version = "0.1.0"
description = ""
authors = ["vslitvinov"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.1.1"
torchvision = "^0.16.1"
tqdm = "^4.66.1"
hydra-core = "^1.3.2"
pandas = "^2.1.4"
mlflow = "^2.9.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.11.0"
isort = "^5.12.0"
pylint = "^3.0.2"
pre-commit = "^3.5.0"
dvc = { extras = ["gdrive"], version = "^3.33.3" }
mypy = "^1.7.1"
flake8-bugbear = "^23.12.2"
flake8-pyproject = "^1.2.3"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88
lines_after_imports = 2

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
